user_management/requirements.txt
aiohttp
cryptography
email-validator
Flask-Assets
Flask-Compress
Flask-CORS
Flask-Login
Flask-Mail
Flask-Migrate==2.7.0
Flask-SQLAlchemy==3.0.5 
Flask-SSE
Flask-WTF
flask==2.3.3
flask_rq2
flask_socketio
gevent
gunicorn
isort
jsmin
phonenumbers
playwright
psycopg2-binary
PyJWT==2.8.0
pyotp
pytz
qrcodegen
qrcode[pil]
raygun4py
redis==5.0.4
redis-py-cluster
rq==1.16.2
WTForms-SQLAlchemy
yapf